# Improve perf of mat4 int8 shaders using 8x8x1 threadgroup