Improved perf of mat4 int8 shaders by using an 8x8x1 threadgroup:

Using device Apple M1 Pro
- Perf of reduce_group_mat4xmat4_int8mm type bfloat dim 32x4128x4096 is 569.543 GFLOPs
- Perf of reduce_mat4xmat4_int8mm type bfloat dim 32x4128x4096 is 369.181 GFLOPs
- Perf of reduce_group_int8mm type bfloat dim 32x4128x4096 is 230.136 GFLOPs
- Perf of reduce_mat4_int8mm type bfloat dim 32x4128x4096 is 150.624 GFLOPs
- Perf of reduce_vec4_int8mm type bfloat dim 32x4128x4096 is 35.5223 GFLOPs
- Perf of naive_int8mm type bfloat dim 32x4128x4096 is 9.01956 GFLOPs